+++ src/main/java/com/dinosaur/foodbowl/domain/user/dao/UserFindDao.java
@@ -1,9 +1,9 @@
 package com.dinosaur.foodbowl.domain.user.dao;
 
-import static com.dinosaur.foodbowl.domain.user.exception.UserErrorCode.USER_NOT_FOUND;
+import static com.dinosaur.foodbowl.global.error.ErrorCode.USER_NOT_FOUND;
 
 import com.dinosaur.foodbowl.domain.user.entity.User;
-import com.dinosaur.foodbowl.domain.user.exception.UserException;
+import com.dinosaur.foodbowl.global.error.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +17,6 @@ public class UserFindDao {
 
   public User findById(final Long id) {
     return memberRepository.findById(id)
-        .orElseThrow(() -> new UserException(id, "userId", USER_NOT_FOUND));
+        .orElseThrow(() -> new BusinessException(id, "userId", USER_NOT_FOUND));
   }
 }
