+++ src/main/java/com/dinosaur/foodbowl/global/error/BusinessException.java
@@ -1,0 +1,21 @@
+package com.dinosaur.foodbowl.global.error;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class BusinessException extends RuntimeException {
+
+  private final Object invalidValue;
+  private final String fieldName;
+  private final HttpStatus httpStatus;
+  private final String message;
+
+  public BusinessException(Object invalidValue, String fieldName, ErrorCode errorCode) {
+    super(errorCode.getMessage());
+    this.invalidValue = invalidValue;
+    this.fieldName = fieldName;
+    this.httpStatus = errorCode.getHttpStatus();
+    this.message = errorCode.getMessage();
+  }
+}
