+++ src/main/java/com/dinosaur/foodbowl/global/error/ErrorCode.java
@@ -1,10 +1,10 @@
-package com.dinosaur.foodbowl.domain.user.exception;
+package com.dinosaur.foodbowl.global.error;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 @Getter
-public enum UserErrorCode {
+public enum ErrorCode {
 
   LOGIN_ID_DUPLICATE("이미 존재하는 로그인 아이디입니다.", HttpStatus.CONFLICT),
   NICKNAME_DUPLICATE("이미 존재하는 닉네임입니다.", HttpStatus.CONFLICT),
@@ -14,7 +14,7 @@ public enum UserErrorCode {
   private final String message;
   private final HttpStatus httpStatus;
 
-  UserErrorCode(String message, HttpStatus httpStatus) {
+  ErrorCode(String message, HttpStatus httpStatus) {
     this.message = message;
     this.httpStatus = httpStatus;
   }
